バッファの更新フラグとメッシュシェーダー頂点作成サイズの修正

+++ BaseProject/Source/Engine/Resource/Data/Mesh/MeshShaderData/MeshShaderData.h
@@ -19,7 +19,7 @@ namespace Engine::Resource
 	struct MeshShaderData
 	{
 		std::vector<Meshlet> meshlets;								// メッシュレットのカタログ
-		std::vector<uint8_t> uniqueVertexIndices;						// 各メッシュレットが使う頂点番号リスト
+		std::vector<uint32_t> uniqueVertexIndices;						// 各メッシュレットが使う頂点番号リスト
 		std::vector<DirectX::MeshletTriangle> primitiveIndices;		// ローカルなインデックス（0 から 63の範囲）
 
 		Graphics::MeshAllocationHandle meshHandle = {};
